Partial graph build.
The program "anaconf" now accepts arguments which are equipment
names. If these arguments are given, only those equipments
configurations are analyzed to rebuild the graph.

If no argument is given, previous behaviour is not changed: the
graph is rebuilt for all equipments.

More optimizations are still to be done, e.g. do not send probes
in case of partial build.