Centralizes all dynamic configuration parameters.
This is a new infrastructure to centralize dynamic configuration
(anything other than critical path to install application or database
parameters) in one place. Everything is now declared in www/lib/libdns.tcl
and not in various files. Scripts www/bin/admpar* are now fully
generic and do not need to be modified. The www/aideparam.htgt is no
longer needed since the www/bin/admparhelp exploits inforamtion from
the new class "config".

Configuration of peripheral applications or utilities should be
simpler in the future.