Don't accept DHCP profile on a static RR without MAC address.
It is inconsistent to configure a DHCP profile on a static RR without
configuring a MAC address. Using DHCP on a specific RR makes sense
only to make an static association MAC-IP. So, if there is no MAC
address, there is no need to have a DHCP profile.

This commits:
- prevents from configuring a DHCP profile without configuring a
    MAC address
- removes (during the upgrade to 1.5 version) all DHCP profiles
    from RR without a MAC address