Reset execute bit on scripts.
Previous commit has reset all file modes to 0644. Some of them
should be 0755. This commit resets them to 0755. By the way, some
files had their modes set to 0755 which is not right: the last 2
commits gave the opportunity to clean all these modes. This is
a direct benefit from removal of $Id$ keywords ;-)