Libdns clean-up.
Introduce new methods for the dnscontext object:

- init-script: must be used in autonomous
    programs (not CGI scripts)
    => modify all utility programs accordingly

- end: closes database descriptor

- result: returns a result (HTML or LaTeX) file
    while performing substitutions, and closes
    database descriptor.
    => modify all CGI scripts accordingly

- make-url (not yet used): builds up an URL from
    a path and an argument array

- writelog: write a line in the log (using the
    now private "log" object)
    => modify all programs accordingly

Remove the lire-correspondant-par-id function, and
use the lire-correspondant function (which is the
old lire-correspondant-par-login renamed).

Remove the lire-utilisateur function, which was
specific to the Topo module, and use the new
lire-correspondant function, which now includes
the specific Topo code. This is a mandatory step
to remove init-topo in the future.