Bug fix: remove primary key on topo.modeq table.
An error was encountered when a user was making
two or more modifications at once in vlans or
equipments (www/bin/admrefmodif).

Those modifications triggered an update of
the topo.modeq table, adding either a modification
for the "_routerdb" or the "_vlan" fake equipments.
The date of the modification was the same, which
violates the integrity constraint.

Since this constraint is not useful, remove it.
It is replaced by an index on eq.