Graphviz attributes are now dynamically configurable.
Graphviz node attributes were previously hardcoded
in www/cgi/l2 and www/cgi/l3 scripts.

The new system:
1- adds a topo.dotattr table which gives,
    for numerically ascending regexp, the
    corresponding graphviz node attributes
2- makes the topo.dotattr table modifiable
    via the adminindex script (and via the
    www/cgi/admref script)
3- makes the www/cgi/l2 and l3 scripts use
    this table

It should be noticed that the item "2-" is
difficult since the display/store-tabular
framework was not ready to integrate a
"false" column for node images.

To display node images in the admref page,
tge topo.dotattr includes a "cache" column
containing, if not empty, the binary for
the image alone. The admref fetches this
"cache" column and, if empty, fills it with
the result of a call to graphviz. This is
done in order to improve performance.