Bugfix in logical interface name while extracting collect point
The logical interface name modification introduced a new format
for collect point extraction. There is a simpler solution :
we use the same format for collect point extraction.
When a logical interface name exists for a L2 node,
we remove the vlan id.
Indeed, topographd, which generate the collect point just needs a
"-" in the vlan id column to generate it correctly.