Add DHCP range import
Add a new keyword "dhcp=" in the "networks" import
operation (see file network.txt for an example).

If this keyword has a non empty value (a domain
name), DHCP is enabled on this network (i.e. static
host mapping will be generated for this network).
    dhcp=example.com

In addition, one or more DHCP ranges may be added:
    dhcp=example.com 172.16.11.100-172.16.11.119