Change column name in dns.zone* tables
Rename the "domaine" column of the dns.zone SQL
table to "name". This is meant to reflect the
real meaning of this information: both a zone
name in menus and the name of the file generated
by the mkzones script.

While doing this, correctly order zones (by address)
in the "modify reverse IPv[46] zones" menu.
Close #71