Add a check with database version.
Since Netmagis upgrades often require an upgrade of the database
schema, include a version number in the database schema, in the
global.config table, and checks this version number in the Netmagis
application.

This needs a new "read-only" flag for each configuration parameter,
and appropriate support in the "application parameter" page.

By the way, fix a bug: when an error occurs very soon at initialization
(e.g. for a CGI script), database access is not always correctly
initialized. The error procedure tried to access user's properties
for the "actions" left menu from the database, resulting in an
internal error.