Convert topo.ifchanges/idrr to equipment name
Adding views to topo.ifchanges is somewhat
difficult, and the added complexity is not
judged useful. So, replace the idrr column
of topo.ifchanges by an "eq" text column
which contains a FQDN.

View is then determined by the location in
the network of the toposendd program.

Warning: these changes have not yet been
fully tested, since it needs a testbed that
I don't have.