Introduce error patterns in tests
Error patterns are an optionnal argument
for test, test-sh and test-cgi.
If this pattern matches, test is considered
as failed. If both normal and error pattern
match, the error pattern takes precedence.

As an example, the t100-init:index-1 checks
that there is no '%..%' hole.

While we are here, remove the line "modify
my password" in www/cgi/index: this hole cannot
be substitued by a known URL when using LDAP
authentication.