Use fully qualified equipement name in i/f changes
When recording interfaces changes, the equipement was previously pointing
to a host in the dns.rr table (via idrr).
Removing this dependancy was **a good thing**.

The unqualified equipment name comes from the topo graph.
But a fully qualified name is needed now since it must match an existing
equipement in the topo.eq table.
This is a temporary fix (see issue #9).