Update example Apache configuration file
The example configuration file was written for
Apache 2.0. The newest version (2.4) obsoletes
some directives, so this commit updates this
example to use either Apache 2.2 or 2.4.
This new configuration cannot provide a simple
file to include for each authentication method.
Thus, the old auth-ldap.conf and auth-pgsql.conf
no longer exist, and all directives are now in
httpd.conf, which must be edited by hand (as
before, anyway).

Issue #116 (not closed here)
Close #122