Support for named aggregated interface on servers
On cisco, to add an interface to a channel-group (bonding or 802.3ad
aggregate), we only provide a number X. This number relates to an
implicitely named "Port-ChannelX" interface:

  interface GigabitEthernet0/13
     channel-group 1 ...

  interface Port-Channel1
    ...

On servers, the aggregated interface can have any name. The modified
parsenet script produces now a channel-group directive with the
name of the aggregated interface (for example "trunk0") :

  interface bge1
    channel-group trunk0 ...

  interface trunk0
    ...

This is recognized by cisco-parse-channel-group when analyzed.