Optional equipment configuration filter in topo
This new feature allows to automatically pipe the network equipment
configuration file through a program of your choice before calling
the parsing script.

If the filter program does not exist for a given equipment, nothing
changes.

Your need to define the configuration variable "filterdir" and
create an executable with the same name as the targeted equipment
type.

Example:
	filterdir = /usr/local/lib/netmagis/topofilters

	/usr/local/lib/netmagis/topofilters/juniper should be
	an executable script which behaves like a pipe

Possible usage:

we modified jrancid to additionaly call
	"show configuration | display inheritance"
and insert the resulting lines as comments in the
file created by rancid :

	##<Inherited>## ... [configuration line] ...
	##<Inherited>## ... [configuration line] ...
	##<Inherited>## ... [configuration line] ...

The following filter extract these lines and create a configuration
file which is then fed to the "analyze" script :

	sed -n '/^##<Inherited>##/s///p' $*